requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ddgs>=6.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
tldextract>=5.1.0
tenacity>=8.2.0
selenium>=4.20.0
undetected-chromedriver>=3.5.5
webdriver-manager>=4.0.1
anthropic>=0.40.0
pydantic>=2.6.0
dnspython>=2.6.0
dateparser>=1.2.0
Django>=5.0
channels>=4.1.0
daphne>=4.1.0
whitenoise>=6.7.0
